pyproject.toml
[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[project]
name = "normalize-mp4"
version = "0.1.0"
description = "Categorize MP4/MKV files into show/season folders based on metadata."
readme = "README.md"
authors = [{ name = "normalize-mp4 contributors" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "ffmpeg-python>=0.2,<0.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[project.scripts]
normalize-mp4 = "normalize_mp4.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/normalize_mp4"]

[tool.hatch.build.targets.sdist]
include = ["src", "README.md", "tests", "pyproject.toml", "flake.nix"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
